requirements.txt
streamlit>=1.24.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.13.0
openpyxl>=3.1.0
supabase>=1.0.3
python-dotenv>=1.0.0 